refactor: 쿠키 클래스명 변경
